# (core): CRC24-Radix64 performance improvement (32 bits in parallel)