--- a/src/lib/Makefile.am
+++ b/src/lib/Makefile.am
@@ -19,6 +19,7 @@ librnp_la_SOURCES	= \
 	crypto/sm2.c \
 	bufgap.c \
 	compress.c \
+	crc24_radix64.c \
 	crypto.c \
 	fingerprint.c \
 	generate-key.c \