--- a/src/tests/rnp_tests.c
+++ b/src/tests/rnp_tests.c
@@ -178,7 +178,8 @@ main(int argc, char *argv[])
       cmocka_unit_test(test_ffi_keygen_json_sub_same_ring),
       cmocka_unit_test(test_ffi_keygen_json_sub_different_ring),
       cmocka_unit_test(test_ffi_keygen_json_sub_pass_required),
-      cmocka_unit_test(test_ffi_detect_key_format)};
+      cmocka_unit_test(test_ffi_detect_key_format),
+      cmocka_unit_test(test_crc24_4byte_slicer)};
 
     /* Each test entry will invoke setup_test before running
      * and teardown_test after running. */