@ronaldtse 2017-10-24
@flowher since this is so awesome and is not dependent on RNP, I wonder if we should at least move it to `libfastcrc`? 😉