@flowher 2017-10-24
@ronaldtse okis, I'll do that. Now thinking about it, it may not work on big-endian processor. But that's quick fix.