@dewyatt 2017-10-24
Why aren't we just using botan's CRC24 (via `pgp_hash_t` or some other way)?