@ni4 2017-10-24
@dewyatt Good question. Botan appears to use the PGP values for CRC. When I was optimizing crc-24 I didn't even think that it may implement it, since there are a lot of CRCs with different polynoms/initial values.