@ni4 2017-10-24
@flowher Did you check the performance for all-zero 1Gb files, or with the random data? For constant values performance may be affected by the cache hits/misses.