@flowher 2017-10-24
@dewyatt it's slower than this