@flowher 2017-10-24
@ni4 I'm using random data in the test. I havn't check all zeros, but that case isn't really interesting