@ronaldtse 2017-10-24
@flowher happy to contribute this to Botan via PR as long as @randombit is happy with it 😉