@flowher 2017-10-25
Code tested on qemu with PowerPC. Works OK