@ronaldtse 2017-10-25
@dewyatt @ni4 if it's OK let's merge it!